pyproject.toml
[tool.poetry]
package-mode = false
name = "MAA-Manager"
version = "0.1.0"
description = ""
authors = ["Nemowang <211108100@smail.nju.edu.cn>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"

[[tool.poetry.source]]
name = "mirrors"
url = "https://mirror.nju.edu.cn/pypi/web/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
